warranty_service/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
clinner==1.12.3
colorlog==3.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.10
django-cors-headers==3.5.0
django-filter==2.4.0
Django-Health-Checks==0.0.1
django-openapi==0.1.1.10
django-rest-swagger==2.2.0
djangorestframework==3.11.2
djangorestframework-jsonapi==3.2.0
et-xmlfile==1.0.1
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
openapi-codec==1.3.2
openpyxl==3.0.5
pandas==1.1.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pythonping==1.0.15
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.0.35
whitenoise==5.2.0
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
